fix(deploy): stop failing deploys that are actually healthy, and two SQL bugs
The 19:21 deploy of f06402b failed with every service working. Docker runs a
container's first health check one interval after start, so a 15s interval left
napkey-core marked "starting" for 15 seconds after it was already answering
/ready. Coolify gave up at 11 seconds -- four seconds before the first check was
even due -- and napkey-web was never started. All three app services now check
every 5s with a 30s start_period, so a slow start no longer reads as a failure.

That deploy's logs also showed two queries failing on a background timer, which
is why nobody noticed them: they warn and the process carries on.

  refreshing operations alerts failed: could not determine data type of $5
  expiring promotional credits failed: inconsistent types deduced for $1

Both are Postgres unable to infer a placeholder's type. jsonb_build_object takes
any type, so $5 had no context; and $1 was used as a uuid in user_id and as text
in ref_id in the same insert, so inference produced a conflict. Casts fix both.

What they cost while broken: the unmatched-payment alert never opened, and
expired trial credit was never swept from dormant wallets, so that liability
stayed on the books. Neither surfaces anywhere a person looks.

pgtest replays canned responses and does not type-check, so no round-trip test
can catch this class -- which is how both reached production untested. The new
test asserts the casts are present in the SQL text instead, and was confirmed to
fail when either cast is removed.

+++ backend/napkey-core/internal/store/operations.go
@@ -22,12 +22,12 @@ type OperationsSummary struct {
 }
 
 type OperationsAlert struct {
-	ID        string         `json:"id"`
-	Type      string         `json:"type"`
-	Severity  string         `json:"severity"`
-	Title     string         `json:"title"`
-	Metadata  map[string]any `json:"metadata"`
-	OpenedAt  time.Time      `json:"openedAt"`
+	ID       string         `json:"id"`
+	Type     string         `json:"type"`
+	Severity string         `json:"severity"`
+	Title    string         `json:"title"`
+	Metadata map[string]any `json:"metadata"`
+	OpenedAt time.Time      `json:"openedAt"`
 }
 
 func (s *Store) ListOpenOperationsAlerts(ctx context.Context, limit int) ([]OperationsAlert, error) {
@@ -74,7 +74,7 @@ func (s *Store) RefreshOperationsAlerts(ctx context.Context) error {
 		}
 		if count > 0 {
 			_, err := s.db.ExecContext(ctx, `INSERT INTO operations_alerts (alert_type, severity, fingerprint, title, metadata)
-				VALUES ($1, $2, $3, $4, jsonb_build_object('count', $5))
+				VALUES ($1, $2, $3, $4, jsonb_build_object('count', $5::bigint))
 				ON CONFLICT (fingerprint) WHERE status = 'open' DO UPDATE SET metadata = excluded.metadata`,
 				item.alertType, item.severity, item.fingerprint, item.title, count)
 			if err != nil {
